CircularDoublyLinkedListIterator: Add initial renaming
A number of questions remain about the proper naming of this. I don't
like the "ContextualEntry" class and I'm not sure how I feel about the
iterator naming. But will have to try it out in the client and see where
the naming fails.

This commit at least confirms the logic works how I expected.

Signed-off-by: Open592 Developer <uss@open592.com>